# docs/swagger.yaml
definitions:
  auth.LoginDto:
    properties:
      email:
        example: makc@mail.ru
        type: string
      password:
        example: sdfsdfs222
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      tokens:
        $ref: '#/definitions/jwt.Tokens'
      user_id:
        example: 1
        type: integer
    type: object
  auth.RegistrationResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  brand.Brand:
    properties:
      description:
        example: бренд одежды...
        type: string
      id:
        example: 2
        type: integer
      img_path:
        example: /static/category/adidas.webp
        type: string
      slug:
        example: adidas
        type: string
      title:
        example: adidas
        type: string
    type: object
  brand.CreateBrandDto:
    properties:
      description:
        example: бренд одежды
        minLength: 10
        type: string
      img_path:
        example: /static/category/adidas.webp
        type: string
      title:
        example: adidas
        minLength: 2
        type: string
    required:
    - title
    type: object
  brand.UpdateBrandDto:
    properties:
      description:
        example: бренд одежды
        minLength: 10
        type: string
      img_path:
        example: /static/category/adidas.webp
        type: string
      title:
        example: adidas
        minLength: 2
        type: string
    type: object
  category.Category:
    properties:
      category_id:
        example: 3
        type: integer
      img_path:
        example: /static/example.webp
        type: string
      level:
        type: integer
      parent_category_id:
        example: 2
        type: integer
      short_title:
        type: string
      slug:
        example: verhnia-odezhda
        type: string
      subcategories:
        items:
          $ref: '#/definitions/category.Category'
        type: array
      title:
        example: Верхняя одежда
        type: string
    type: object
  category.CreateCategoryDto:
    properties:
      img_path:
        example: /static/example.webp
        type: string
      parent_category_id:
        example: 4
        type: integer
      short_title:
        example: Верхняя одежда
        minLength: 3
        type: string
      title:
        example: Мужская Верхняя одежда
        minLength: 3
        type: string
    required:
    - short_title
    - title
    type: object
  category.UpdateCategoryDto:
    properties:
      img_path:
        example: /static/example.webp
        type: string
      short_title:
        example: Верхняя одежда
        minLength: 3
        type: string
      title:
        example: Мужская Верхняя одежда
        minLength: 3
        type: string
    type: object
  exception.AppErr:
    properties:
      message:
        example: some error
        type: string
      status:
        example: 404
        type: integer
    type: object
  exception.ValidationError:
    properties:
      errors:
        items:
          $ref: '#/definitions/exception.validationErrorItem'
        type: array
      status:
        example: 400
        type: integer
    type: object
  exception.validationErrorItem:
    properties:
      key:
        example: email
        type: string
      message:
        example: email is invalid
        type: string
    type: object
  jwt.Tokens:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  user.CreateUserDto:
    properties:
      email:
        example: makc@mail.ru
        type: string
      password:
        example: sdfsdfs222
        maxLength: 100
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact: {}
paths:
  /api/auth/registration:
    post:
      consumes:
      - application/json
      description: Registation by body arguments
      parameters:
      - description: Registation user with body dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegistrationResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Registation user
      tags:
      - auth
  /api/brand/:
    get:
      consumes:
      - application/json
      description: Get all brands
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/brand.Brand'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get all brands
      tags:
      - brand
    post:
      consumes:
      - application/json
      description: Create brand with dto
      parameters:
      - description: Create brand dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/brand.CreateBrandDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create brand
      tags:
      - brand
  /api/brand/:id:
    patch:
      consumes:
      - application/json
      description: Update brand
      parameters:
      - description: Update brand dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/brand.UpdateBrandDto'
      - description: id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Update brand
      tags:
      - brand
  /api/brand/:title:
    get:
      consumes:
      - application/json
      description: Get brand by title
      parameters:
      - description: title parameter
        in: path
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/brand.Brand'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get brand by title
      tags:
      - brand
  /api/category/:
    get:
      consumes:
      - application/json
      description: Get all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/category.Category'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get all categories
      tags:
      - category
    post:
      consumes:
      - application/json
      description: Create category with dto
      parameters:
      - description: Create category dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/category.CreateCategoryDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create category
      tags:
      - category
  /api/category/:id:
    patch:
      consumes:
      - application/json
      description: Update category
      parameters:
      - description: Update category dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/category.UpdateCategoryDto'
      - description: id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Update category
      tags:
      - category
  /api/category/with-sub/:slug:
    get:
      consumes:
      - application/json
      description: Get by slug with sub categories
      parameters:
      - description: slug parameter
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/category.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get by slug with sub categories
      tags:
      - category
...
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login to an account with account data
      parameters:
      - description: login in account
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/auth.LoginDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Login
      tags:
      - auth
  /api/user/refresh-token:
    get:
      consumes:
      - application/json
      description: Refresh tokens by cookies refresh_token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/exception.AppErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Refresh tokens
      tags:
      - auth
swagger: "2.0"